replace double with loss in loss.cu and adam_optimizer.cu

+++ HugeCTR/src/loss.cu
@@ -174,37 +174,37 @@ BinaryCrossEntropyLoss::BinaryCrossEntropyLoss(
 // Suppose we use one thread to calculate one sample
 __global__ void BinaryCrossEntropy_Kernel(float *input, const float *label, float *bce_loss,
                                           int scaler, int batch_size, float rterm) {
-  const double MIN_ = 1e-7;
+  const float MIN_ = 1e-7;
   const float MIN_X = -18.f;
   int tid = threadIdx.x;
   extern __shared__ float loss_s[];
   loss_s[tid] = 0.0f;
 
   float x, y;
-  double val;
+  float val;
 
   for (int i = tid; i < batch_size; i += blockDim.x) {
 
     x = input[i];
-    double exp_neg_x = 0.;
+    float exp_neg_x = 0.f;
     if(x > MIN_X){
-      exp_neg_x = exp((double)-x);
+      exp_neg_x = exp(-x);
       val = 1.0f / (1.0f + exp_neg_x);
     }
     else{
-      val = 0.;
+      val = 0.f;
     }
 
     y = label[i];
-    val = clip(val, MIN_, 1.0 - MIN_);
+    val = clip(val, MIN_, 1.0f - MIN_);
     loss_s[tid] += y * log(val) + (1.0f - y) * log(1.0f - val);
 
     // grad
-    if(x > MIN_X && val < 1.0 - MIN_ && val > MIN_){
-      input[i] = -1.0 * val * (y - val) * exp_neg_x / (1.0 - val) / batch_size * scaler;
+    if(x > MIN_X && val < 1.0f - MIN_ && val > MIN_){
+      input[i] = -1.0f * val * (y - val) * exp_neg_x / (1.0f - val) / batch_size * scaler;
     }
     else{
-      input[i] = 0;
+      input[i] = 0.f;
     }
     
 //     if(isnan(input[i]) && (!isnan(x))){
